Rollback to FP16 GEMM pipeline (~107 TPS)
Restore the original FP16 I/O + FP16 accumulator + split-K path that
achieved ~107 TPS prefill. All FP32 end-to-end experiments were slower.

Restored:
- int4_gemm_adreno.cl: half8 accumulators, half I/O
- input_transpose.cl: read_imageh/write_imageh
- repack_kai_to_adreno.cl: half scales
- blas_kernels.cpp: CL_HALF_FLOAT images, uint16_t SVM buffers
- float_tensor.cpp: host NEON FP32<->FP16 + input pointer caching

+++ nntrainer/tensor/cl_operations/cl_kernels/repack_kai_to_adreno.cl
@@ -9,7 +9,7 @@
 __attribute__((qcom_reqd_sub_group_size("full"))) kernel void
 repack_kai_to_adreno(const __global uchar* kai_packed_data,
     __global ushort* weights,
-    __global float* scales,
+    __global half* scales,
     const int N, const int K, const int rhs_packed_stride, const int quantization_group_size) {
 
     const int k_id = get_global_id(0);
@@ -74,7 +74,7 @@ repack_kai_to_adreno(const __global uchar* kai_packed_data,
         int scale_offset_bytes = NR * (K_aligned / 2 + 4);
         const __global float* scale_src = (const __global float*)(kai_packed_data + base + scale_offset_bytes);
         for (int sub_n = 0; sub_n < NR; sub_n++) {
-            scales[n_id * NR + sub_n] = scale_src[sub_n] * 16.0f;
+            scales[n_id * NR + sub_n] = (half)(scale_src[sub_n] * 16.0f);
         }
     }
 }
